% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avgROISignal.R
\name{avgROISignal}
\alias{avgROISignal}
\title{Get ROI Average Signals}
\usage{
avgROISignal(labels, labeledBoldMat, providePlot = TRUE)
}
\arguments{
\item{labels}{ROI labels}

\item{labeledBoldMat}{Labeled BOLD signal matrix}

\item{providePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
Average ROI signal matrix
}
\description{
Compute the mean time singal for each ROI using ROI labels.
}

