% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collectNuisParams.R
\name{collectNuisParams}
\alias{collectNuisParams}
\title{Identify Nuisance Parameters in Timeseries}
\usage{
collectNuisParams(boldMat, seg, mask, goodtimes, badtimes, providePlot = TRUE)
}
\arguments{
\item{boldMat}{BOLD Time Series matrix}

\item{seg}{Tissue segmentation image that identifies: CSF, gray and white matter}

\item{mask}{mask for BOLD image (3D)}

\item{goodtimes}{list of interger index of good timepoints}

\item{badtimes}{list of interger index of bad timepoints}

\item{providePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
list of outputs containing:
\itemize{
\item{ctxVox:}{ Index of cortex voxels}
\item{ctxMean:}{ Mean of cortex voxel signals}
\item{wmVox:}{ Index of white matter}
\item{wmMean:}{ Mean of white matter signals}
\item{csfVox:}{ Index of CSF}
\item{csfMean:}{ Mean of CSF signals}
\item{globalMean:}{ Mean of Global signals}
\item{compcorNuis:}{ Physiological nuisance parameters}
\item{tissueNuis:}{ Tissue nuisance parameters}
\item{tissueDeriv:}{ Derivative of tissue nuisance singals}
}
}
\description{
Collect the nuisance parameters in BOLD global signals to regress out.
Possible nuisance parameters:
\itemize{
\item{detrended motion parameters, their squares, and the derivatives of both}
\item{mean signal in white Matter and its derivative}
\item{mean signal in CSF and its derivative}
\item{physiologocial noise estimated via compcor()}
}
}

