% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corrMatROI.R
\name{corrMatROI}
\alias{corrMatROI}
\title{Compute Correlation matrix for ROI signals}
\usage{
corrMatROI(roiMat, labels)
}
\arguments{
\item{roiMat}{ROI average signal matrix}

\item{labels}{A set of labels identifying anatomical regions of interest}
}
\value{
ROI signal correlation matrix
}
\description{
Use the mean time signal for each ROI and find the correlation between
each ROI signals
}

