% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityGraph.R
\name{densityGraph}
\alias{densityGraph}
\title{Create Network Density Graph}
\usage{
densityGraph(corMat, pts, density = 0.1, providePlot = TRUE)
}
\arguments{
\item{corMat}{ROI correlation matrix}

\item{pts}{labels identifying anatomical regions of interest}

\item{%}{of edges to retain in the correlation matrix, default is 0.1 (10%)}

\item{Boolean}{determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
adjacency matrix
}
\description{
Binarize ROI correlation matrix and create an adjacency matrix. The
adjacency matrix is then used to create network density graph from graph
operations. Default density is 0.1
}

