% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detrendTimeseries.R
\name{detrendTimeseries}
\alias{detrendTimeseries}
\title{Demand and Detrend the Time-series Data}
\usage{
detrendTimeseries(img, mask, goodtimes, badtimes, providePlot = TRUE)
}
\arguments{
\item{img}{BOLD fMRI time series image}

\item{mask}{for image (3D)}

\item{goodtimes}{list of interger index of good timepoints}

\item{badtimes}{list of interger index of bad timepoints}

\item{providePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\description{
Demean and detrend BOLD time series while excluding bad timepoints in BOLD
images
}

