% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportStats.R
\name{exportStats}
\alias{exportStats}
\title{Output Descriptive Statistics from BOLD processing}
\usage{
exportStats(outdir, prefix)
}
\arguments{
\item{outdir}{string indicating the output directory}

\item{prefix}{string indicating the prefix of output filename}
}
\description{
Provide all the plots associated with each steps of BOLD processing.
The plots are cached from function call
}

