% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterFrequency.R
\name{filterFrequency}
\alias{filterFrequency}
\title{Filter and Smooth BOLD Timeseries Frequency}
\usage{
filterFrequency(img, mask, boldMat, ctxVox, goodtimes, badtimes,
  lowFreq = 0.009, highFreq = 0.08, smoothFactor = 6,
  providePlot = TRUE)
}
\arguments{
\item{img}{BOLD fMRI time series image}

\item{mask}{mask for BOLD image (3D)}

\item{boldMat}{BOLD time series matrix}

\item{ctxVox}{list of integer index of mean cortex signal vortex}

\item{goodtimes}{list of interger index of good timepoints}

\item{badtimes}{list of interger index of bad timepoints}

\item{lowFreq}{lower bound of frequency for filtering, default is 0.009 Hz}

\item{highFreq}{upper bound of frequency for filtering, default is 0.08 Hz}

\item{smoothFactor}{frequency smooth factor ,default FWHM is 6.0mm}

\item{providePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
list of output containing:
\itemize{
\item{img}{Processed 4D fMRI image}
\item{boldMat}{Processed BOLD time series matrix}
}
}
\description{
Frequency filtering get rid of non-evenly sampled time-series data. After
filtering out bad timepoints, spatial smoothing is applied.The Default
frequency range for filter is 0.009 Hz - 0.08 Hz 0.009 Hz - 0.08 Hz, default
smooth factor is Gaussian kernel with FWHM=6.0mm
}

