% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findBadTimepoints.R
\name{findBadTimepoints}
\alias{findBadTimepoints}
\title{Identify Bad Timpoints in BOLD Image}
\usage{
findBadTimepoints(img, meanDisplacement, threshold = 0.1,
  providePlot = TRUE)
}
\arguments{
\item{img}{BOLD fMRI time series image}

\item{meanDisplacement}{mean framewise displacement of BOLD image}

\item{threshold}{threshold of framewise displacement, default is 0.1mm}

\item{providePlotBoolean}{determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
list of outputs containing:
\itemize{
\item{good:}{ Index of good timepoints}
\item{bad:}{ Index of bad timepoints}
}
}
\description{
Use Motion Registration Parameters to identify bad timepoints in time series.
The bad timepoints are deteremined by mean framewise displacement.
}

