% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSteadyState.R
\name{findSteadyState}
\alias{findSteadyState}
\title{Remove pre Steady State Timepoints}
\usage{
findSteadyState(img, mask, mocoParams = NULL, preSteady = 10,
  providePlot = TRUE)
}
\arguments{
\item{img}{BOLD fMRI time series image}

\item{mask}{mask for BOLD image (3D)}

\item{mocoParams}{if moco parameters data includes non steady state timepoints, the data
should be provided. Otherwise, no input is required}

\item{preSteady}{time points to exclude, default is set to 10 seconds}

\item{providePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
list of outputs containing:
\itemize{
\item{ss_img}{Steady State BOLD image}
\item{OPTIONAL: ss_MOCOParams}{Steady State MocoParameters}
}
}
\description{
Identify initial time points that occur before magnetization steady state
in BOLD fMRI and remove pre steady state timepoints. Default is set to
excluding the first 10 seconds of the data.
}
\examples{
fmri <- antsImageRead(getANTsRData("rsbold"))
msk <- antsImageRead(getANTsRData("rsboldmask"))
ssResult <- rsBOLDSteadyState(fmri, msk)
}

