% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDisplacement.R
\name{getDisplacement}
\alias{getDisplacement}
\title{Calculate Framewise Displacement of BOLD Image}
\usage{
getDisplacement(img, mask, param, providePlot = TRUE)
}
\arguments{
\item{img}{motion corrected BOLD fMRI time series image}

\item{mask}{mask for image(3D)}

\item{param}{MOCO registraion parameters data frame}

\item{providePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
a dataframe consisting mean displacement and max displacement
}
\description{
Measures the average displacement of voxels betweeen consecutive time points
}

