% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphMeasure.R
\name{graphMeasure}
\alias{graphMeasure}
\title{Examine Graph-metrics from Node and/or Global Measures}
\usage{
graphMeasure(adjMat, globalMeasure = TRUE, providePlot = TRUE)
}
\arguments{
\item{adjMat}{Network adjacency matrix}

\item{globalMeasure}{Boolean determining whether the entire network in a single measure
should bed summarized}

\item{providePlot}{Boolean determining whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
list of node measures, containing:
\itemize{
\item{Degree：}{ Number of links connected to a node}
\item{localClustCoeff：}{ Local neighborhood connectivity}
\item{pathLen:}{ Mean shortest distance between one node and all others}
\item{localEff:}{ Closeness of nodes in a neighborhood}
\item{pageRank: }{ Google page rank measure}
\item{(OPTIONAL) globalClustCoeff： }{ Closeness of nodes to all other nodes}
\item{(OPTIONAL) globalEff： }{ Global neighborhood connectivity}
}
}
\description{
Obtain graph-metrics from node measures to examine various properties of
functional network
}

