% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressNuisParams.R
\name{regressNuisParams}
\alias{regressNuisParams}
\title{Regress out Nuisance Parameters in Time Series}
\usage{
regressNuisParams(boldMat, param, NuisList, goodtimes, providePlot = TRUE)
}
\arguments{
\item{boldMat}{BOLD time series matrix}

\item{param}{MOCO registration parameters}

\item{NuisList}{BOLD Nuisance Parameter list generated by rsBOLDCollectNuisParams()}

\item{goodtimes}{list of interger index of good timepoints}

\item{providePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
BOLD time series matrix after regression
}
\description{
Regress out BOLD data Nuisance Parameters, the Parameters are identified in
rsBOLDCollectNuisParams().
}

