% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/systemROI.R
\name{systemROI}
\alias{systemROI}
\title{Comparing ROI signal between systems}
\usage{
systemROI(labels, labeledBoldMat, roiMat, pts, providePlot = TRUE)
}
\arguments{
\item{labels}{labels}

\item{labeledBoldMat}{labeled BOLD time series matrix}

\item{roiMat}{average ROI matrix}

\item{pts}{labels identifying anatomical regions of interest}

\item{providePlotprovidePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
list of output containing:
\itemize{
\item{sysMatMean:}{ System BOLD signal mean matrix}
\item{sysMatSD:}{ System BOLD signal standard deviation matrix}
}
}
\description{
Given ROI system labels, provide the mean and the standard deviation for
BOLD signal within system. Provides insight on how the systems are working
together
}

