% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avgROISignal.R
\name{avgROISignal}
\alias{avgROISignal}
\title{Get ROI Average Signals}
\usage{
avgROISignal(boldMat, mask, labelImg, goodtimes, providePlot = TRUE)
}
\arguments{
\item{boldMat}{BOLD timer series matrix}

\item{mask}{mask for BOLD image (3D)}

\item{labelImg}{ROI labels image}

\item{goodtimes}{list of interger index of good timepoints}

\item{providePlot}{Boolean determine whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
ROI matrix
}
\description{
Get ROI Average Signals
}

