% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defineROI.R
\name{defineROI}
\alias{defineROI}
\title{Defining ROI of functional network}
\usage{
defineROI(mask, pts, radius = 5)
}
\arguments{
\item{pts}{labels identifying anatomical regions of interest}

\item{maks}{mask for BOLD image (3D)}

\item{rqdius}{radius of each ROI sphere, default is 5.0mm}
}
\value{
ROI labels image
}
\description{
A set of ROI is needed for creating network, define ROIs that contain many
voxels and represent regions for with priori knowledge regarding the
functional network to which each region belongs. Default radius of ROI
sphere is 5.0mm
}

