% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findTemporalMean.R
\name{findTemporalMean}
\alias{findTemporalMean}
\title{Calculate the Temporal Mean of BOLD image}
\usage{
findTemporalMean(img, providePlot = FALSE)
}
\arguments{
\item{img}{bold input 4D image}

\item{providePlot}{boolean for plotting, default is FALSE}
}
\value{
antsImage of temporal mean
}
\description{
Find the Mean of BOLD data Time Series, this is done by calculating
the mean of each voxel at all timepoints
}
\examples{
fmri <- antsImageRead(getANTsRData("rsbold"))
tempMean <- rsBOLDTemporalMean(fmri)
}

