% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphMeasure.R
\name{graphMeasure}
\alias{graphMeasure}
\title{Examine Graph-metrics from Node and/or Global Measures}
\usage{
graphMeasure(adjMat, globalMeasure = TRUE, providePlot = TRUE)
}
\arguments{
\item{adjMat}{Network adjacency matrix}

\item{globalMeasure}{Boolean determining whether the entire network in a single measure
should bed summarized}

\item{providePlot}{Boolean determining whether descriptive statistics should be provided
if TRUE, the plot will be cached and can be called}
}
\value{
list of node measures, containing:
\itemize{
\item{Degree}{the number of connections that include the node}
\item{localClustCoeff}{local neighborhood connectivity}
\item{pathLen}{mean shortest distance between one node and all others}
\item{localEff}{"closeness" off nodes in a neighborhood}
\item{pageRank}{Google page rank measure}
\item(OPTIONAL: globalClustCoeff){ closeness of nodes to all other nodes}
\item{OPTIONAL: globalEff}{ global neighborhood connectivity}
}
}
\description{
Obtain graph-metrics from node measures to examine various properties of
functional network
}

